makeDocument bugfix
Electron had updated its dialog.showSaveDialog() method to be async, now returning an object rather than a string as was the case before. Accessing the object for the path string solved this problem. Jszip also had some major changes between versions 2 and 3, so I've switched back to the previous version for now.